refactor(mcp): restructure server and minor fixes
