+++ app/bioblend_server/mcp_context.py
@@ -1,0 +1,3 @@
+from contextvars import ContextVar
+
+current_api_key_server: ContextVar[str] = ContextVar("current_api_key_server", default=None)
